refactor: one catalog client for every provider, and no per-model cloud lookup
Three structural changes, all following from the same finding: the catalog was
already the source of truth, but the code around it did not treat it as one.

1. The app-wide catalog loader stops pretending to be an OpenRouter feature.
   fetch/cache/warm/refreshCatalog and the module-level _memCache lived inside
   OpenRouterCatalogResolver, yet writeAllModelsCache had exactly ONE caller in
   the codebase and every catalog read — context windows, reasoning capability,
   tokenParam, routeVariant — depended on that class having run. The name hid
   the dependency and made "why is there only one resolver?" unanswerable.
   Extracted verbatim to providers/catalog-client.ts.

2. The per-provider resolver interface is gone. Name resolution was never
   provider-specific logic — it is one aggregators[] lookup, which the catalog
   populates for 18 providers. A registry of classes to perform one table lookup
   per provider is a registry of one thing repeated. resolveExternalId replaces
   it, and is verified byte-identical to the old resolver across all 1543
   distinct catalog inputs (every modelId, alias, and externalId): 1543/1543.

   The generic path also STRICTLY DOMINATES the bespoke antigravity lookup added
   earlier today, so lookupFamilyDefaultVariant is deleted: it matches all three
   cases that had a routeVariant.isDefault AND resolves gemini-2.5-pro and
   gemini-3-flash-preview, which the isDefault-only lookup missed.

   Resolution now runs for EVERY provider, not just OpenRouter — the gate was
   why each provider grew its own translation. mm@minimax-m2.5 now sends
   MiniMax-M2.5 (their api_official spelling, and their API is case-sensitive);
   together-ai@glm-5 sends zai-org/GLM-5; kimi-k3 sends k3. A model the catalog
   does not know — a local GGUF, a custom endpoint id — passes through.

3. The per-model cloud fallback is deleted. resolveCatalogContextWindow issued
   one HTTP request per unknown model (?search=<id>) against the SAME cloud that
   had just served the bulk catalog: an N+1 that returns the same answer N
   round-trips later. It was also chasing data that does not exist — of 150 slim
   entries with no contextWindow, ~148 are not chat models at all (embeddings,
   ASR, TTS, video); ~2 are plausibly text. A missing window now renders as
   unknown, which keeps the gap visible as a models-index issue instead of
   papering over it on every launch. Three call sites removed.

The two sources are now cleanly split: the catalog owns model IDENTITY (same for
everyone), a provider's live endpoint owns ENTITLEMENT (which subset this key may
use, and the window for this tier — Kimi k3 is 1M only on Allegretto+). Joined on
model id, not chained as fallbacks.

Also fixes a regression this refactor exposed. With gemini-3.1-pro-preview now
resolving to the served gemini-3.1-pro-high, the request reaches Antigravity —
which 400s on that variant (a known backend quirk: it is in the account's served
set AND marked the catalog family default, yet never generates). A 400 is
terminal, so the chain ABORTED where it previously 404'd and fell through. The
fallback handler's 400-advances-chain rule is a message allowlist that did not
cover Antigravity's generic "Request contains an invalid argument", so an
Antigravity 400 now advances a bare-name chain. Scoped to that provider on
purpose: a blanket rule would mask real payload bugs elsewhere. Verified live —
"Antigravity failed (HTTP 400), trying next provider... OpenRouter succeeded".

NOT done, deliberately: moving Antigravity's fetchAvailableModels onto a
modelDiscovery descriptor. That descriptor is GET + auth headers +
openai-models-list; Antigravity needs POST, a body carrying a projectId from a
prior setupAntigravityUser round-trip, and a models-dict-keys response. Forcing
it would add method/body-builder/provider-state to an abstraction that serves
its two current providers cleanly. Widening the format union with no second
consumer is speculative generality.

Verified: typecheck clean, build clean, 1684 tests across 99 files, 0 failures.
Test delta from baseline is fully accounted for: 22 tests in the deleted
openrouter.test.ts plus 1 null-resolver test whose defensive branch no longer
exists (its real scenario, missing cache + fetch_failed, is already covered).

Backend asks recorded in models-index TASK_model_behavior_metadata_gaps.md:
drop isDefault from the non-functional gemini-3.1-pro-high, guarantee
contextWindow on chat models, and stop leaking embedding/ASR/TTS/video models
into the chat catalog.


Crafted with agentic harness Magus (https://github.com/MadAppGang/magus)
